[MIG] product_medical: Migration to 14.0
