+++ product_medical/data/medical_data.xml
@@ -21,31 +21,4 @@
     <record id="medical_class_7" model="medical.class">
         <field name="name">CE Class III</field>
     </record>
-    <record id="in_vitro_1" model="in.vitro.diagnostic">
-        <field name="name">InVitroDiagnostika-A</field>
-    </record>
-    <record id="in_vitro_2" model="in.vitro.diagnostic">
-        <field name="name">InVitroDiagnostika-B</field>
-    </record>
-    <record id="in_vitro_3" model="in.vitro.diagnostic">
-        <field name="name">InVitroDiagnostika-C</field>
-    </record>
-    <record id="in_vitro_4" model="in.vitro.diagnostic">
-        <field name="name">InVitroDiagnostika-D</field>
-    </record>
-    <record id="category_d" model="medicine.category">
-        <field name="name">Drug-D</field>
-    </record>
-    <record id="category_e" model="medicine.category">
-        <field name="name">Drug-E</field>
-    </record>
-    <record id="ppe_category_1" model="ppe.category">
-        <field name="name">|</field>
-    </record>
-    <record id="ppe_category_2" model="ppe.category">
-        <field name="name">||</field>
-    </record>
-    <record id="ppe_category_3" model="ppe.category">
-        <field name="name">|||</field>
-    </record>
 </odoo>
