+++ product_medical/data/ppe_category_data.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo noupdate="1">
+
+    <record id="ppe_category_1" model="ppe.category">
+        <field name="name">|</field>
+    </record>
+    <record id="ppe_category_2" model="ppe.category">
+        <field name="name">||</field>
+    </record>
+    <record id="ppe_category_3" model="ppe.category">
+        <field name="name">|||</field>
+    </record>
+
+</odoo>
