+++ product_medical/models/__init__.py
@@ -1,2 +1,5 @@
 from . import product_template
 from . import medical_class
+from . import in_vitro_diagnostic
+from . import medical_category
+from . import ppe_category
