+++ product_medical/models/in_vitro_diagnostic.py
@@ -1,0 +1,16 @@
+from odoo import fields, models
+
+
+class InVitroDiagnostics(models.Model):
+    _name = "in.vitro.diagnostic"
+    _description = "In vitro diagnostics"
+
+    name = fields.Char(string="Name", required=True, translate=True)
+
+    _sql_constraints = [
+        (
+            "name_uniq",
+            "unique(name)",
+            "Diagnostic type already exists",
+        )
+    ]
