+++ product_medical/models/medical_class.py
@@ -17,48 +17,3 @@ class MedicalClass(models.Model):
             "Medical class already exists",
         )
     ]
-
-
-class InVitroDiagnostics(models.Model):
-    _name = "in.vitro.diagnostic"
-    _description = "In vitro diagnostics"
-
-    name = fields.Char(string="Name", required=True, translate=True)
-
-    _sql_constraints = [
-        (
-            "name_uniq",
-            "unique(name)",
-            "Diagnostic type already exists",
-        )
-    ]
-
-
-class MedicineCategory(models.Model):
-    _name = "medicine.category"
-    _description = "Medicine Category"
-
-    name = fields.Char(string="Name", required=True, translate=True)
-
-    _sql_constraints = [
-        (
-            "name_uniq",
-            "unique(name)",
-            "Medicine category already exists",
-        )
-    ]
-
-
-class PPECategory(models.Model):
-    _name = "ppe.category"
-    _description = "PPE category"
-
-    name = fields.Char(string="Name", required=True)
-
-    _sql_constraints = [
-        (
-            "name_uniq",
-            "unique(name)",
-            "PPE category already exists",
-        )
-    ]
