+++ product_medical/models/ppe_category.py
@@ -1,0 +1,16 @@
+from odoo import fields, models
+
+
+class PPECategory(models.Model):
+    _name = "ppe.category"
+    _description = "PPE category"
+
+    name = fields.Char(string="Name", required=True)
+
+    _sql_constraints = [
+        (
+            "name_uniq",
+            "unique(name)",
+            "PPE category already exists",
+        )
+    ]
