+++ product_medical/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* Iryna Vyshnevska, Camptocamp
+* `Trobz <https://trobz.com>`_:
+  * Dung Tran <dungtd@trobz.com>
