+++ product_medical/readme/CREDITS.rst
@@ -1,0 +1,5 @@
+The development of this module has been financially supported by:
+
+* Akretion <https://akretion.com>
+* La Base <https://labase.coop>
+* Camptocamp
